refactor: useJsonStyles from Reponse -> hooks/useCodeMirrorJsonStyles.ts
